Socioeconomic data column names

A string list containing values for the column names of socioeconomic data. See pwt10::pwt10.0.

Usage

socioecon_cols

Format

A string list with 14 entries.

isocode_colname

The name of a metadata column containing the 3-letter isocodes of countries.

year_colname

The name of a metadata column containing values for the year of the observation. Lower case.

rgdpe_colname

The name of the column containing data for Expenditure-side real GDP at chained PPPs (in million 2017 USD).

rgdpo_colname

The name of the column containing data for Output-side real GDP at chained PPPs (in million 2017 USD).

rgdpna_colname

The name of the column containing data for Real GDP at constant 2017 national prices (in million 2017 USD).

emp_colname

The name of the column containing data for Number of persons engaged (in millions).

avh_colname

The name of the column containing data for Average annual hours worked by persons engaged.

hc_colname

The name of the column containing data for Human capital index, based on years of schooling and returns to education; see Human capital in PWT9.

rnna_colname

The name of the column containing data for Capital stock at constant 2017 national prices (in million 2017 USD).

rkna_colname

The name of the column containing data for Capital services at constant 2017 national prices (2017 = 1).

K_colname

A more representative name for rnna_colname.

Kserv_colname

A more representative name for rkna_colname.

L_colname

The name of the column containing data for the total number of hours worked in a given year. See PFUWorkflow::get_L_K_GDP_data.

Ladj_colname

The name of the column containing data for the total number of hours worked adjusted by the human capital index. See PFUWorkflow::get_L_K_GDP_data.

Examples

socioecon_cols
#> $isocode_colname
#> [1] "isocode"
#> 
#> $year_colname
#> [1] "year"
#> 
#> $rgdpe_colname
#> [1] "rgdpe"
#> 
#> $rgdpo_colname
#> [1] "rgdpo"
#> 
#> $rgdpna_colname
#> [1] "rgdpna"
#> 
#> $emp_colname
#> [1] "emp"
#> 
#> $avh_colname
#> [1] "avh"
#> 
#> $hc_colname
#> [1] "hc"
#> 
#> $rnna_colname
#> [1] "rnna"
#> 
#> $rkna_colname
#> [1] "rkna"
#> 
#> $K_colname
#> [1] "K"
#> 
#> $Kserv_colname
#> [1] "Kserv"
#> 
#> $L_colname
#> [1] "L"
#> 
#> $Ladj_colname
#> [1] "Ladj"
#>