Targets cache target names

A string list containing names of targets targets.

Usage

target_names

Format

A string list with 48 entries.

countries

The countries to be analyzed in this pipeline.

additional_exemplar_countries

The additional exemplar countries from which allocations and efficiencies can be obtained.

alloc_and_eff_countries

The allocation and efficiency countries target in the targets cache, giving the countries from which final-to-useful allocations and efficiencies can be drawn for final-to-useful analyses. This is a superset of countries to be analyzed.

years

The years to be analyzed.

iea_data_path

The path to the file containing IEA extended energy balance data.

country_concordance_path

The path to the country concordance file.

phi_constants_path

The path to exergy-to-energy ration (phi) constants.

ceda_data_folder

The path to the folder containing CEDA data.

machine_data_path

The path to the folder containing machine data excel files.

exemplar_table_path

The path to the exemplar table.

fu_analysis_folder

The path to a folder containing final-to-useful analyses for various countries.

report_source_folders

The path to a folder containing reports to be run as the last step of the pipeline.

report_dest_folder

The path to a folder containing reports from the pipeline.

pipeline_caches_folder

The path to a folder containing pipeline caches for each release.

pipeline_releases_folder

The path to a folder containing releases of targets.

release

A boolean telling whether a release was requested.

iea_data

The IEA data for countries.

country_concordance_table

A data frame containing concordance information which maps full country names to custom 3 letter codes.

ceda_data

The name of the data frame containing all CEDA temperature data read from ceda_data_folder.

all_machine_data

A data frame containing eta_fu values.

machine_data

A filtered version of all_machine_data containing information for only alloc_and_eff_countries.

socio_econ_data

A data frame containing socioeconomic data, supplied by get_L_K_GDP_data().

balanced_before

A boolean indicating whether the iea_data are balanced before any further analysis. They usually are not, so this value is typically FALSE.

balanced_iea_data

A balanced version of iea_data.

balanced_after

Same as balanced_before, only for after balancing. This should always be be TRUE.

ok_to_proceed

A boolean telling whether we can continue the pipeline. This target errors if balanced_after is not TRUE.

specified

A data frame containing specified IEA data.

psut_final

A data frame containing specified in a PSUT format.

exemplar_lists

A data frame of lists of exemplar countries for each country in countries, and maybe more.

phi_constants

A data frame of constant values for exergy-to-energy ratio (phi).

incomplete_allocation_tables

A data frame of final-to-useful allocation tables, one for each country. These allocation tables may be incomplete.

completed_allocation_tables

A data frame of completed final-to-useful allocation tables.

completed_efficiency_tables

A data frame of completed final-to-useful efficiency tables.

completed_phi_u_tables

A data frame of completed useful-stage exergy-to-energy ratios.

cmats

A data frame containing CompletedAllocationTables in matrix form.

eta_fu_phi_u_vecs

A data frame containing final-to-useful efficiency vectors and useful exergy-to-energy ratios.

eta_fu_vecs

A data frame containing final-to-useful efficiency vectors.

phi_u_vecs

A data frame containing useful exergy-to-energy ratio vectors.

phi_pf_vecs

A data frame containing primary and final exergy-to-energy ratio vectors.

phi_vecs

A data frame containing exergy-to-energy ratio vectors.

psut_useful

A data frame containing PSUT matrices up to the useful stage.

psut

A data frame containing PSUT matrices up to the useful stage and with exergy.

allocation_graphs

A data frame containing final-to-useful allocation graphs.

non_stationary_allocation_graphs

A data frame containing final-to-useful allocation graphs, for non-stationary data only.

efficiency_graphs

A data frame containing final-to-useful efficiency graphs.

phi_graphs

A data frame containing exergy-to-energy ratio (phi) graphs.

release_psut

A target that does a release of the psut target. Contains the name of the target or a message saying that a release was not requested.

store_cache

A target that stores the targets cache. Contains the name of the target or a message saying that a release was not requested.

Examples

target_names
#> $countries
#> [1] "Countries"
#> 
#> $additinoal_exemplar_countries
#> [1] "AdditionalExemplarCountries"
#> 
#> $alloc_and_eff_countries
#> [1] "AllocAndEffCountries"
#> 
#> $years
#> [1] "Years"
#> 
#> $iea_data_path
#> [1] "IEADataPath"
#> 
#> $country_concordance_path
#> [1] "CountryConcordancePath"
#> 
#> $phi_constants_path
#> [1] "PhiConstantsPath"
#> 
#> $ceda_data_folder
#> [1] "CEDADataFolder"
#> 
#> $machine_data_path
#> [1] "MachineDataPath"
#> 
#> $exemplar_table_path
#> [1] "ExemplarTablePath"
#> 
#> $fu_analysis_folder
#> [1] "FUAnalysisFolder"
#> 
#> $report_source_folders
#> [1] "ReportsSourceFolders"
#> 
#> $report_dest_folder
#> [1] "ReportsDestFolder"
#> 
#> $pipeline_caches_folder
#> [1] "PipelineCachesFolder"
#> 
#> $pipeline_releases_folder
#> [1] "PipelineReleasesFolder"
#> 
#> $release
#> [1] "Release"
#> 
#> $iea_data
#> [1] "IEAData"
#> 
#> $country_concordance_table
#> [1] "CountryConcordanceTable"
#> 
#> $ceda_data
#> [1] "CEDAData"
#> 
#> $all_machine_data
#> [1] "AllMachineData"
#> 
#> $machine_data
#> [1] "MachineData"
#> 
#> $socio_econ_data
#> [1] "SocioEconData"
#> 
#> $balanced_before
#> [1] "BalancedBefore"
#> 
#> $balanced_iea_data
#> [1] "BalancedIEAData"
#> 
#> $balanced_after
#> [1] "BalancedAfter"
#> 
#> $ok_to_proceed
#> [1] "OKToProceed"
#> 
#> $specified
#> [1] "Specified"
#> 
#> $psut_final
#> [1] "PSUTFinal"
#> 
#> $exemplar_lists
#> [1] "ExemplarLists"
#> 
#> $phi_constants
#> [1] "PhiConstants"
#> 
#> $incomplete_allocation_tables
#> [1] "IncompleteAllocationTables"
#> 
#> $completed_allocation_tables
#> [1] "CompletedAllocationTables"
#> 
#> $completed_efficiency_tables
#> [1] "CompletedEfficiencyTables"
#> 
#> $completed_phi_u_tables
#> [1] "CompletedPhiuTables"
#> 
#> $cmats
#> [1] "Cmats"
#> 
#> $eta_fu_phi_u_vecs
#> [1] "EtafuPhiuvecs"
#> 
#> $eta_fu_vecs
#> [1] "Etafuvecs"
#> 
#> $phi_u_vecs
#> [1] "Phiuvecs"
#> 
#> $phi_pf_vecs
#> [1] "Phipfvecs"
#> 
#> $phi_vecs
#> [1] "Phivecs"
#> 
#> $psut_useful
#> [1] "PSUTUseful"
#> 
#> $psut
#> [1] "PSUT"
#> 
#> $allocation_graphs
#> [1] "AllocationGraphs"
#> 
#> $non_stationary_allocation_graphs
#> [1] "NonStationaryAllocationGraphs"
#> 
#> $efficiency_graphs
#> [1] "EfficiencyGraphs"
#> 
#> $phi_graphs
#> [1] "PhiGraphs"
#> 
#> $release_psut
#> [1] "ReleasePSUT"
#> 
#> $store_cache
#> [1] "StoreCache"
#>